server/requirements.txt
fastapi
uvicorn
python-dotenv
authlib
httpx
starlette-session
redis
motor
python-jose[cryptography]